+++ backend/src/services/TicketServices/UpdateTicketService.ts
@@ -293,7 +293,7 @@ const UpdateTicketService = async ({
         !isGroup &&
         !ticket.contact.disableBot &&
         !justClose &&
-        ticket.whatsapp?.complationMessage.trim() &&
+        ticket.whatsapp?.complationMessage?.trim() &&
         ticket.whatsapp.status === "CONNECTED"
       ) {
         const body = formatBody(